feat: Add an interactive JSON viewer
Intelligently collapse large arrays and objects (except array items),
and arrays that are direct children of array items.

The react-json-tree component is only loaded after the page is loaded in
the browser. This means that clients without JavaScript (or with slow
network connnections) can still view the site without waiting for the
JavaScript bundle to load. The user can freely switch between the
interactive viewer and the raw JSON by toggling the checkbox.

As an aid to users, also display the size in kB of the raw JSON
(formatted).