refactor: Make TableOfContents a functional component
Move the scroll event listener into a separate ActiveSection component.
This component takes a render function and a list of objects that
contain an id property as props, then calls the render function,
passing in as an argument the object whose id matches the ID of the
active page section.