Switch to React-Static from Gatsby
Gatsby forces you to use GraphQL for pre-processing data, and is
generally inflexible. There's an overview here:
https://medium.com/@tannerlinsley/%EF%B8%8F-introducing-react-static-a-progressive-static-site-framework-for-react-3470d2a51ebc
that covers the main reasons why I prefer React-Static, but it basically
boils down to: I didn't want to deal with GraphQL and custom plugins to
pre-process the documentation data so the docs page doesn't need to ship
>4 MB of JS, including a Markdown parser, two copies of the data
(in HTML and JSON), doesn't need to parse that JSON multiple times between
search-and-replacing the domain name, etc.

This commit only deals with moving to React-Static; the docs page is
still dog-slow, but the next commits will improve it considerably.