Refactor everything
This commit touches almost every file in the codebase.
There should be a lot of speed improvements, especially on the
Documentation page.
A lot less code is being shipped to the browser,
and the random bugs on the Documentation table of contents should be gone.
Visually, it's almost identical to the previous version.